% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete.R
\name{hue_delete_group}
\alias{hue_delete_group}
\title{Delete group}
\usage{
hue_delete_group(id)
}
\arguments{
\item{id}{If numeric, numeric id of group. If character, name of group You can check id and names with \code{hue_get_groups_names()}}
}
\value{
Silently returns http response from the bridge, but mostly used for its side effects.
}
\description{
Delete group
}
