% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete.R
\name{hue_delete_light}
\alias{hue_delete_light}
\title{Delete light}
\usage{
hue_delete_light(id)
}
\arguments{
\item{id}{If numeric, numeric id of light. If charachter, name of light. You can check id and names with \code{hue_get_lights_names()}}
}
\value{
Silently returns http response from the bridge, but mostly used for its side effects.
}
\description{
Delete light
}
