% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_group.R
\name{hue_get_group_lights}
\alias{hue_get_group_lights}
\title{Get state of given group}
\usage{
hue_get_group_lights(id)
}
\arguments{
\item{id}{If numeric, numeric id of group. If character, name of group. You can check id and names with \code{hue_get_groups_names()}}
}
\value{
An integer vector with id of lights included in a group.
}
\description{
Get state of given group
}
