% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_group_cards.R
\name{hue_mod_group_card_server}
\alias{hue_mod_group_card_server}
\title{Server function for shiny app}
\usage{
hue_mod_group_card_server(id)
}
\arguments{
\item{id}{Module id}
}
\value{

}
\description{
Server function for shiny app
}
