% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_group_cards.R
\name{hue_mod_group_card_ui}
\alias{hue_mod_group_card_ui}
\title{UI function for shiny app}
\usage{
hue_mod_group_card_ui(
  id,
  group,
  onoff,
  brightness,
  temperature,
  min_temperature,
  max_temperature
)
}
\arguments{
\item{id}{Module id}

\item{group}{Group id}
}
\description{
UI function for shiny app
}
