% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hue_output_group_id}
\alias{hue_output_group_id}
\title{Check if given id is valid}
\usage{
hue_output_group_id(id)
}
\arguments{
\item{id}{If numeric, numeric id of light. If character, name of light. You can check id and names with \code{hue_get_lights_names()}}
}
\value{
Always return an integer corresponding to a given light.
}
\description{
Check if given id is valid
}
