% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hue_output_id}
\alias{hue_output_id}
\title{Title}
\usage{
hue_output_id(id)
}
\arguments{
\item{id}{If numeric, numeric id of light. If charachter, name of light. You can check id and names with \code{hue_get_lights_names()}}
}
\value{
Always return an integer corresponding to a given light.
}
\description{
Title
}
