% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_state.R
\name{hue_set_group_state}
\alias{hue_set_group_state}
\title{Change state of Hue groups}
\usage{
hue_set_group_state(id, params)
}
\arguments{
\item{id}{If numeric, numeric id of group. If character, name of group. You can check id and names with \code{hue_get_groups_names()}}

\item{params}{A named list. For a full list of available parameters, check https://developers.meethue.com/develop/hue-api/groupds-api/}
}
\description{
Change state of Hue groups
}
