% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_light_temperature.R
\name{hue_set_group_temperature}
\alias{hue_set_group_temperature}
\title{Set temperature of group}
\usage{
hue_set_group_temperature(id, temperature, by = 10)
}
\arguments{
\item{id}{If numeric, numeric id of group. If character, name of group. You can check id and names with \code{hue_get_groups_names()}}

\item{temperature}{A numeric value, or one of either "warmer" (or "+") or "colder" (or "-"). The warmest value available is usually 500, the coldest 153.}

\item{by}{Numeric, defaults to 10. Ignored if temperature is numeric. Otherwise determines the size of the increment/decrement.}
}
\value{
Silently returns http response from the bridge, but mostly used for its side effects.
}
\description{
Set temperature of group
}
