% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_light_brightness.R
\name{hue_set_light_brightness}
\alias{hue_set_light_brightness}
\title{Set brightness of light}
\usage{
hue_set_light_brightness(id, brightness, by = 10)
}
\arguments{
\item{id}{If numeric, numeric id of light. If character, name of light. You can check id and names with \code{hue_get_lights_names()}}

\item{brightness}{A numeric value, or one of either "brighter" (or "+") or "darker" (or "-"). The highest value is 254, the lowest 1.}

\item{by}{Numeric, defaults to 10. Ignored if brightness is numeric. Otherwise determines the size of the increment/decrement.}
}
\value{
Silently returns http response from the bridge, but mostly used for its side effects.
}
\description{
Set brightness of light
}
