% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hue_set_colour.R
\name{hue_set_light_colour}
\alias{hue_set_light_colour}
\title{Set colour of a Hue light with relevant capabilities}
\usage{
hue_set_light_colour(id, colour, transition_time = 0.4)
}
\arguments{
\item{id}{If numeric, numeric id of light. If character, name of light. You can check id and names with \code{hue_get_lights_names()}}

\item{colour}{A colour name as listed by \code{colours()} or a hexadecimal colour string.}

\item{transition_time}{Duration of the transition in seconds. Defaults to 0.4 seconds.}
}
\description{
N.B. Colour accuracy may not be perfect. See the official documentation for details, and consider using custom parameters with \code{hue_set_light_state()} if this function does not achieve what you expect. https://developers.meethue.com/develop/application-design-guidance/color-conversion-formulas-rgb-to-xy-and-back/
}
\examples{

if (interactive()) {
  hue_set_light_colour(id = 11, colour = "blue")
}

}
