% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_state.R
\name{hue_set_light_state}
\alias{hue_set_light_state}
\title{Change state of Hue lights}
\usage{
hue_set_light_state(id, params)
}
\arguments{
\item{id}{If numeric, numeric id of light. If character, name of light. You can check id and names with \code{hue_get_lights_names()}}

\item{params}{A named list. For a full list of available parameters, check https://developers.meethue.com/develop/hue-api/lights-api/}
}
\description{
Change state of Hue lights
}
