% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{hue_settings}
\alias{hue_settings}
\title{Set IP and username to interact with the api}
\usage{
hue_settings(ip = NULL, username = NULL)
}
\arguments{
\item{ip}{A character vector of length one. The IP of your hue bridge, e.g. "192.168.0.1"}

\item{username}{A username generated to interact with the API. For details, https://developers.meethue.com/develop/get-started-2/}
}
\description{
Set IP and username to interact with the api
}
