% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{hue_shiny_controller}
\alias{hue_shiny_controller}
\title{Interactive interface to control lights}
\usage{
hue_shiny_controller(options = list())
}
\value{

}
\description{
options Options to be passed to the shiny app. See \code{?shiny::shinyApp()} for details.
}
