% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turn_on_off.R
\name{hue_turn_light_on}
\alias{hue_turn_light_on}
\title{Turn on light}
\usage{
hue_turn_light_on(id)
}
\arguments{
\item{id}{If numeric, numeric id of light. If character, name of light. You can check id and names with \code{hue_get_lights_names()}}
}
\value{
Silently returns http response from the bridge, but mostly used for its side effects.
}
\description{
Turn on light
}
