% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turn_on_off.R
\name{hue_turn_off}
\alias{hue_turn_off}
\title{Turn off light}
\usage{
hue_turn_off(id)
}
\arguments{
\item{id}{If numeric, numeric id of light. If charachter, name of light. You can check id and names with \code{hue_get_lights_names()}}
}
\value{
Silently returns http response from the bridge, but mostly used for its side effects.
}
\description{
Turn off light
}
