% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huecontroller-package.R
\docType{package}
\name{huecontroller-package}
\alias{huecontroller}
\alias{huecontroller-package}
\title{huecontroller: Control Philips Hue lights using the R programming language}
\description{
Control Philips Hue lights using the R programming language
}
\seealso{
Useful links:
\itemize{
  \item \url{https://giocomai.github.io/huecontroller}
}

}
\author{
\strong{Maintainer}: Giorgio Comai \email{g@giorgiocomai.eu} (\href{https://orcid.org/0000-0002-0515-9542}{ORCID})

}
\keyword{internal}
