% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_light_cards.R
\name{hue_mod_light_card_ui}
\alias{hue_mod_light_card_ui}
\title{UI function for shiny app}
\usage{
hue_mod_light_card_ui(
  id,
  light,
  onoff,
  brightness,
  temperature,
  min_temperature,
  max_temperature
)
}
\arguments{
\item{id}{Module id}

\item{light}{Light id}
}
\description{
UI function for shiny app
}
